// frontend/api/model/userPublicDto.ts
/**
 * API Documentation
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RoleDto } from './roleDto';


export interface UserPublicDto { 
    id?: string;
    username?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    status?: UserPublicDto.StatusEnum;
    roles?: Set<RoleDto>;
    createDate?: string;
    updateDate?: string;
}
export namespace UserPublicDto {
    export type StatusEnum = 'ACTIVE' | 'INACTIVE' | 'DELETED';
    export const StatusEnum = {
        Active: 'ACTIVE' as StatusEnum,
        Inactive: 'INACTIVE' as StatusEnum,
        Deleted: 'DELETED' as StatusEnum
    };
}